Move Batch4 StringBuilder tests -> Batch1
