+++ Tests/Batch4/Batch4.csproj
@@ -135,7 +135,6 @@
     <Compile Include="Collections\Generic\GenericDictionaryTests.cs" />
     <Compile Include="Serialization\JsonTests.cs" />
     <Compile Include="Text\RegularExpressions\RegexTests.cs" />
-    <Compile Include="Text\StringBuilderTests.cs" />
     <Compile Include="Collections\Generic\ListTests.cs" />
     <Compile Include="Reflection\TypeSystemLanguageSupportTests.cs" />
     <Compile Include="Reflection\TypeSystemTests.cs" />
