+++ Tests/Runner/Batch4/Bridge.Test/bridge.clienttest.batch4.runner.js
@@ -1640,27 +1640,6 @@ Bridge.assembly("Bridge.Test.Bridge.ClientTest.Batch4", function ($asm, globals)
             QUnit.test("RegexTests - ExecWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.RegexTests.ExecWorks);
             QUnit.test("RegexTests - LastIndexWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.RegexTests.LastIndexWorks);
             QUnit.test("RegexTests - TestWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.RegexTests.TestWorks);
-            QUnit.test("StringBuilderTests - TypePropertiesAreCorrect", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.TypePropertiesAreCorrect);
-            QUnit.test("StringBuilderTests - DefaultConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.DefaultConstructorWorks);
-            QUnit.test("StringBuilderTests - ConstructorWithCapacityWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.ConstructorWithCapacityWorks);
-            QUnit.test("StringBuilderTests - InitialTextConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.InitialTextConstructorWorks);
-            QUnit.test("StringBuilderTests - InitialTextConstructorWithCapacityWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.InitialTextConstructorWithCapacityWorks);
-            QUnit.test("StringBuilderTests - AppendBoolWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendBoolWorks);
-            QUnit.test("StringBuilderTests - AppendCharWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendCharWorks);
-            QUnit.test("StringBuilderTests - AppendIntWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendIntWorks);
-            QUnit.test("StringBuilderTests - AppendDoubleWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendDoubleWorks);
-            QUnit.test("StringBuilderTests - AppendObjectWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendObjectWorks);
-            QUnit.test("StringBuilderTests - AppendStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendStringWorks);
-            QUnit.test("StringBuilderTests - AppendLineWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendLineWorks);
-            QUnit.test("StringBuilderTests - AppendLineBoolWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendLineBoolWorks);
-            QUnit.test("StringBuilderTests - AppendLineCharWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendLineCharWorks);
-            QUnit.test("StringBuilderTests - AppendLineIntWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendLineIntWorks);
-            QUnit.test("StringBuilderTests - AppendLineDoubleWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendLineDoubleWorks);
-            QUnit.test("StringBuilderTests - AppendLineObjectWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendLineObjectWorks);
-            QUnit.test("StringBuilderTests - AppendLineStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.AppendLineStringWorks);
-            QUnit.test("StringBuilderTests - ClearWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.ClearWorks);
-            QUnit.test("StringBuilderTests - ToStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.ToStringWorks);
-            QUnit.test("StringBuilderTests - LengthPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests.LengthPropertyWorks);
             QUnit.test("AsyncTests - AsyncVoid", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.AsyncTests.AsyncVoid);
             QUnit.test("AsyncTests - AsyncTask", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.AsyncTests.AsyncTask);
             QUnit.test("AsyncTests - AsyncTaskBodyThrowsException", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.AsyncTests.AsyncTaskBodyThrowsException);
@@ -10102,131 +10081,6 @@ Bridge.assembly("Bridge.Test.Bridge.ClientTest.Batch4", function ($asm, globals)
                 var $t;
                 if (this.context == null) {
                     this.context = ($t = new Bridge.Test.Runtime.FixtureContext(), $t.Project = "Batch4", $t.ClassName = "Bridge.ClientTest.Batch4.StopwatchTests", $t.File = "Batch4\\StopwatchTests.cs", $t);
-                }
-                return this.context;
-            }
-        }
-    });
-
-    Bridge.define("Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests", {
-        inherits: [Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests)],
-        statics: {
-            methods: {
-                TypePropertiesAreCorrect: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "TypePropertiesAreCorrect()", $t.Line = "17", $t));
-                    t.Fixture.TypePropertiesAreCorrect();
-                },
-                DefaultConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DefaultConstructorWorks()", $t.Line = "26", $t));
-                    t.Fixture.DefaultConstructorWorks();
-                },
-                ConstructorWithCapacityWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ConstructorWithCapacityWorks()", $t.Line = "34", $t));
-                    t.Fixture.ConstructorWithCapacityWorks();
-                },
-                InitialTextConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "InitialTextConstructorWorks()", $t.Line = "42", $t));
-                    t.Fixture.InitialTextConstructorWorks();
-                },
-                InitialTextConstructorWithCapacityWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "InitialTextConstructorWithCapacityWorks()", $t.Line = "50", $t));
-                    t.Fixture.InitialTextConstructorWithCapacityWorks();
-                },
-                AppendBoolWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendBoolWorks()", $t.Line = "68", $t));
-                    t.Fixture.AppendBoolWorks();
-                },
-                AppendCharWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendCharWorks()", $t.Line = "77", $t));
-                    t.Fixture.AppendCharWorks();
-                },
-                AppendIntWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendIntWorks()", $t.Line = "86", $t));
-                    t.Fixture.AppendIntWorks();
-                },
-                AppendDoubleWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendDoubleWorks()", $t.Line = "95", $t));
-                    t.Fixture.AppendDoubleWorks();
-                },
-                AppendObjectWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendObjectWorks()", $t.Line = "104", $t));
-                    t.Fixture.AppendObjectWorks();
-                },
-                AppendStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendStringWorks()", $t.Line = "113", $t));
-                    t.Fixture.AppendStringWorks();
-                },
-                AppendLineWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendLineWorks()", $t.Line = "122", $t));
-                    t.Fixture.AppendLineWorks();
-                },
-                AppendLineBoolWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendLineBoolWorks()", $t.Line = "131", $t));
-                    t.Fixture.AppendLineBoolWorks();
-                },
-                AppendLineCharWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendLineCharWorks()", $t.Line = "140", $t));
-                    t.Fixture.AppendLineCharWorks();
-                },
-                AppendLineIntWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendLineIntWorks()", $t.Line = "149", $t));
-                    t.Fixture.AppendLineIntWorks();
-                },
-                AppendLineDoubleWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendLineDoubleWorks()", $t.Line = "158", $t));
-                    t.Fixture.AppendLineDoubleWorks();
-                },
-                AppendLineObjectWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendLineObjectWorks()", $t.Line = "167", $t));
-                    t.Fixture.AppendLineObjectWorks();
-                },
-                AppendLineStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AppendLineStringWorks()", $t.Line = "176", $t));
-                    t.Fixture.AppendLineStringWorks();
-                },
-                ClearWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ClearWorks()", $t.Line = "185", $t));
-                    t.Fixture.ClearWorks();
-                },
-                ToStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToStringWorks()", $t.Line = "194", $t));
-                    t.Fixture.ToStringWorks();
-                },
-                LengthPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.Text.StringBuilderTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.StringBuilderTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "LengthPropertyWorks()", $t.Line = "202", $t));
-                    t.Fixture.LengthPropertyWorks();
-                }
-            }
-        },
-        fields: {
-            context: null
-        },
-        methods: {
-            GetContext: function () {
-                var $t;
-                if (this.context == null) {
-                    this.context = ($t = new Bridge.Test.Runtime.FixtureContext(), $t.Project = "Batch4", $t.ClassName = "Bridge.ClientTest.Batch4.Text.StringBuilderTests", $t.File = "Batch4\\Text\\StringBuilderTests.cs", $t);
                 }
                 return this.context;
             }
