+++ Tests/Runner/Batch4/code.js
@@ -21474,139 +21474,6 @@ Bridge.assembly("Bridge.ClientTest.Batch4", {"Bridge.ClientTest.Batch4.Reflectio
         }
     });
 
-    Bridge.define("Bridge.ClientTest.Batch4.Text.StringBuilderTests", {
-        methods: {
-            TypePropertiesAreCorrect: function () {
-                var sb = new System.Text.StringBuilder();
-                Bridge.Test.NUnit.Assert.AreEqual("System.Text.StringBuilder", Bridge.Reflection.getTypeFullName(System.Text.StringBuilder));
-                Bridge.Test.NUnit.Assert.True(Bridge.Reflection.isClass(System.Text.StringBuilder));
-                Bridge.Test.NUnit.Assert.True(Bridge.hasValue(sb));
-            },
-            DefaultConstructorWorks: function () {
-                var sb = new System.Text.StringBuilder();
-                Bridge.Test.NUnit.Assert.AreEqual("", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(0, sb.getLength(), "Length");
-            },
-            ConstructorWithCapacityWorks: function () {
-                var sb = new System.Text.StringBuilder("", 55);
-                Bridge.Test.NUnit.Assert.AreEqual("", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(0, sb.getLength(), "Length");
-            },
-            InitialTextConstructorWorks: function () {
-                var sb = new System.Text.StringBuilder("some text");
-                Bridge.Test.NUnit.Assert.AreEqual("some text", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(9, sb.getLength(), "Length");
-            },
-            InitialTextConstructorWithCapacityWorks: function () {
-                var sb = new System.Text.StringBuilder("some text", 55);
-                Bridge.Test.NUnit.Assert.AreEqual("some text", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(9, sb.getLength(), "Length");
-            },
-            AppendBoolWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.append(true), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|true", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(5, sb.getLength(), "Length");
-            },
-            AppendCharWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.append(String.fromCharCode(99)), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|c", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(2, sb.getLength(), "Length");
-            },
-            AppendIntWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.append(123), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|123", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(4, sb.getLength(), "Length");
-            },
-            AppendDoubleWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.append(123.0), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|123", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(4, sb.getLength(), "Length");
-            },
-            AppendObjectWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.append(new Bridge.ClientTest.Batch4.Text.StringBuilderTests.SomeClass()), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|some text", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(10, sb.getLength(), "Length");
-            },
-            AppendStringWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.append("some text"), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|some text", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(10, sb.getLength(), "Length");
-            },
-            AppendLineWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.appendLine(), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|\r\n", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(3, sb.getLength(), "Length");
-            },
-            AppendLineBoolWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.appendLine(System.Boolean.toString((true))), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|True\r\n", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(7, sb.getLength(), "Length");
-            },
-            AppendLineCharWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.appendLine(String.fromCharCode((99))), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|c\r\n", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(4, sb.getLength(), "Length");
-            },
-            AppendLineIntWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.appendLine((123).toString()), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|123\r\n", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(6, sb.getLength(), "Length");
-            },
-            AppendLineDoubleWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.appendLine(System.Double.format((123.0))), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|123\r\n", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(6, sb.getLength(), "Length");
-            },
-            AppendLineObjectWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.appendLine((new Bridge.ClientTest.Batch4.Text.StringBuilderTests.SomeClass()).toString()), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|some text\r\n", sb.toString(), "Length");
-                Bridge.Test.NUnit.Assert.AreEqual(12, sb.getLength(), "Length");
-            },
-            AppendLineStringWorks: function () {
-                var sb = new System.Text.StringBuilder("|");
-                Bridge.Test.NUnit.Assert.True(Bridge.referenceEquals(sb.appendLine("some text"), sb));
-                Bridge.Test.NUnit.Assert.AreEqual("|some text\r\n", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(12, sb.getLength(), "Length");
-            },
-            ClearWorks: function () {
-                var sb = new System.Text.StringBuilder("some text");
-                sb.clear();
-                Bridge.Test.NUnit.Assert.AreEqual("", sb.toString(), "Text");
-                Bridge.Test.NUnit.Assert.AreEqual(0, sb.getLength(), "Length");
-            },
-            ToStringWorks: function () {
-                // Yes, this is tested by every other test as well. Included for completeness only
-                var sb = new System.Text.StringBuilder("some text");
-                Bridge.Test.NUnit.Assert.AreEqual("some text", sb.toString());
-            },
-            LengthPropertyWorks: function () {
-                // Yes, this is tested by every other test as well. Included for completeness only
-                var sb = new System.Text.StringBuilder("some text");
-                Bridge.Test.NUnit.Assert.AreEqual(9, sb.getLength());
-            }
-        }
-    });
-
-    Bridge.define("Bridge.ClientTest.Batch4.Text.StringBuilderTests.SomeClass", {
-        methods: {
-            toString: function () {
-                return "some text";
-            }
-        }
-    });
-
     Bridge.define("Bridge.ClientTest.Batch4.Threading.AsyncTests", {
         methods: {
             AsyncVoid: function () {
